Fix syctl-p glob issue
Dir.glob ran in the compile phase, which is too early:  New sysctl settings won't
have a file on disk at this point for the glob to match, so won't be
loaded when the command is actually executed at the end of the chef
run.

Switch to using the shell to glob filenames at the time the command runs
instead.  This is the same command line Ubuntu uses to load sysctl
settings at boot.